+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/manager/delegates/FlipperActionNotifier.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.bridge.api.manager.delegates
+
+import kotlinx.coroutines.flow.Flow
+
+interface FlipperActionNotifier {
+    fun getActionFlow(): Flow<Unit>
+
+    fun notifyAboutAction()
+}