+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/utils/Constants.kt
@@ -1,13 +1,15 @@
 package com.flipperdevices.bridge.api.utils
 
+import com.flipperdevices.bridge.api.model.FlipperVersionInformation
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
 object Constants {
     const val DEVICENAME_PREFIX = "Flipper"
     const val KEYS_DEFAULT_STORAGE = "/any/"
-    const val API_SUPPORTED_VERSION = 0.3f
+    val API_SUPPORTED_VERSION = FlipperVersionInformation(majorVersion = 0, minorVersion = 3)
     const val LAGS_FLIPPER_DETECT_TIMEOUT_MS = 30 * 1000L // 30 seconds
+    const val FLIPPER_AUTODISCONNECT_TIMEOUT_MS = 5 * 60 * 1000L // 5 minutes
 
     object GenericService {
         val SERVICE_UUID: UUID = UUID.fromString("00001800-0000-1000-8000-00805f9b34fb")