+++ components/bridge/service/impl/src/main/java/com/flipperdevices/bridge/service/impl/FlipperServiceApiImpl.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.bridge.service.impl
 
+import android.app.Application
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.lifecycle.LifecycleOwner
@@ -8,6 +9,8 @@ import com.flipperdevices.bridge.api.error.FlipperServiceErrorListener
 import com.flipperdevices.bridge.api.manager.FlipperBleManager
 import com.flipperdevices.bridge.impl.manager.FlipperBleManagerImpl
 import com.flipperdevices.bridge.service.api.FlipperServiceApi
+import com.flipperdevices.bridge.service.impl.delegate.FlipperActionNotifierImpl
+import com.flipperdevices.bridge.service.impl.delegate.FlipperAutoDisconnect
 import com.flipperdevices.bridge.service.impl.delegate.FlipperLagsDetectorImpl
 import com.flipperdevices.bridge.service.impl.delegate.FlipperSafeConnectWrapper
 import com.flipperdevices.bridge.service.impl.di.FlipperServiceComponent
@@ -27,7 +30,7 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.plus
 
-class FlipperServiceApiImpl constructor(
+class FlipperServiceApiImpl(
     context: Context,
     lifecycleOwner: LifecycleOwner,
     serviceErrorListener: FlipperServiceErrorListener
@@ -46,15 +49,35 @@ class FlipperServiceApiImpl constructor(
     @Inject
     lateinit var sentryApi: Shake2ReportApi
 
+    @Inject
+    lateinit var application: Application
+
     init {
         ComponentHolder.component<FlipperServiceComponent>().inject(this)
     }
 
     private val scope = lifecycleOwner.lifecycleScope + Dispatchers.Default
     private val connectionStateProvider = WeakConnectionStateProvider(scope)
-    private val lagsDetector = FlipperLagsDetectorImpl(scope, this, connectionStateProvider)
+    private val flipperActionNotifier = FlipperActionNotifierImpl(scope)
+    private val flipperAutoDisconnect = FlipperAutoDisconnect(
+        scope, flipperActionNotifier, this, application
+    )
+    private val lagsDetector = FlipperLagsDetectorImpl(
+        scope = scope,
+        serviceApi = this,
+        connectionStateProvider = connectionStateProvider,
+        flipperActionNotifier = flipperActionNotifier
+    )
     private val bleManager: FlipperBleManager = FlipperBleManagerImpl(
-        context, settingsStore, scope, serviceErrorListener, lagsDetector, sentryApi, metricApi
+        context,
+        settingsStore,
+        pairSettingsStore,
+        scope,
+        serviceErrorListener,
+        lagsDetector,
+        flipperActionNotifier,
+        sentryApi,
+        metricApi
     ).apply {
         connectionStateProvider.initialize(this.connectionInformationApi)
     }
@@ -72,6 +95,7 @@ class FlipperServiceApiImpl constructor(
             error { "Service api already inited" }
             return
         }
+        flipperAutoDisconnect.init()
         info { "Internal init and try connect" }
         var deviceId: String? = null
         scope.launch(Dispatchers.Default) {