+++ components/updater/screen/src/main/java/com/flipperdevices/updater/screen/composable/ComposableUpdateFailedContent.kt
@@ -55,7 +55,7 @@ private fun ComposableUpdateFailedContent(
 @Composable
 fun ComposableFailedUploadContent() {
     ComposableUpdateFailedContent(
-        imageId = DesignSystem.drawable.pic_flipper_upload_failed,
+        imageId = R.drawable.ic_flipper_upload_failed,
         titleId = R.string.update_screen_failed_flipper_title,
         descriptionId = R.string.update_screen_failed_flipper_desc
     ) {