+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
-# 1.1.5 - In progress
+# 1.1.6 - In progress
 
 - [Feature] Compose Theme
+- [Feature] Add black flipper support (update and info screen)
+- [Feature] Compose Theme(module `core:ui:theme`, module `archive`)
+- [Feature] Auto disconnect after 5 minutes inactivity
 - [BUGFIX] Lock portrait orientation
 - [BUGFIX] Update card not shown when flipper not connected
 - [BUGFIX] Static URL on dynamic link docs
+- [BUGFIX] Fix unsupported version comparison
+
+# 1.1.5 - HotFix
+
+- [HOTFIX] Move "Fix unsupported version comparison" to 1.1.4 release
 
 # 1.1.4